fix: restructure as per PR review, don't pick up conftest
